backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
anthropic==0.40.0
pdfplumber==0.11.4
python-docx==1.1.2
httpx==0.28.0
